requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-allauth>=0.54.0
python-dotenv>=1.0.0
requests>=2.31.0
social-auth-app-django>=5.2.0
